Pikmin holding tools now also trigger `on_pikmin_land`
This was not included in the original commit as I was unsure whether it was a bug or not. However, after noticing that the latch check was the reason `on_pikmin_land` was triggered at all, I assumed that it is meant to trigger while holing a tool as well.