Fixed "on_pikmin_land" event not triggering with Pikmin using the "impact" attack method or holding tools. (#27)
- Fixed "on_pikmin_land" event not triggering with Pikmin using the "impact" attack method.
- Pikmin holding tools now also trigger "on_pikmin_land".

This was not included in the original commit as I was unsure whether it was a bug or not. However, after noticing that the latch check was the reason "on_pikmin_land" was triggered at all, I assumed that it is meant to trigger while holding a tool as well.